FINALLY implemented ability to see all orders from customer
this code is incredibly sloppy but it works. there's some logic in the client which shouldn't be there but I tried using the backend and despite it working on the backend, it didnt work on the front end. I have spent hours on this. I need to get it working somehow even if it's messy